put randn/u in sep. files; get func. for mu/sd
